Fix #2319 caching of constant Field expressions
We only cache Field expressions that do not contain variables.  For
instance:

  Field<Project>(p => p.Name)
  Field<Project>(p => p.Name.Suffix("foo"))
  Field<Project>(p => p.Metadata["foo"])

are all cachable since they only contain constant expressions.

whereas

  Field<Project>(p => p.Name.Suffix(myVariable))
  Field<Project>(p => p.Metadata[myVariable])

are not cachable since we can't evaluate the value of `myVariable` to be
used as a cache key.

However, there was a bug which this commit fixes where we would cache a
variable expression that ended with a constant.  For instance:

  Field<Project>(p => p.Metadata[myVariable].Suffix("foo"))

This expression was incorrectly cached because HasConstantExpressionVisitor
only concerned itself with the last method call: `Suffix("foo")`.

HasConstantExpressionVisitor now visits each method call in the expression
and keeps an audit trail.  If any variables are found within the tree, then
the expression is flagged as not cachable.